% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIResult.R
\docType{methods}
\name{addSerie}
\alias{addSerie}
\alias{addSerie,TDIResult-method}
\title{Add column(s) to series for `TDIResult`}
\usage{
addSerie(obj, x)

\S4method{addSerie}{TDIResult}(obj, x)
}
\arguments{
\item{obj}{An object of class `TDIResult`.}

\item{x}{Time-series to add to series.}
}
\value{
Object `TDIResult` with updated series.
}
\description{
Add additional columns to slot series for object of class `TDIResult`.
}
\seealso{
Other TDIResult generics: 
\code{\link{TDIResult-class}},
\code{\link{getSeries}()},
\code{\link{getSymbol}()},
\code{\link{setSeries}()}
}
\concept{TDIResult generics}
