% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIResult.R
\name{addSeries}
\alias{addSeries}
\title{Add time-series to slot}
\usage{
addSeries(obj, x)
}
\arguments{
\item{obj}{An object of [TDIResult-class].}

\item{x}{Time-series to add to series.}
}
\value{
Object TDIResult.
}
\description{
Add time-series to slot
}
\seealso{
Other TDIResult generics: 
\code{\link{TDIResult-class}},
\code{\link{getSeries}()},
\code{\link{getSymbol}()},
\code{\link{setSeries}()}
}
\concept{TDIResult generics}
