% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIDriver.R
\name{apiConnect}
\alias{apiConnect}
\title{Provide API connection
Retrieves information on objects of classes inheriting from [TDIObject-class].}
\usage{
apiConnect(obj, ...)
}
\arguments{
\item{obj}{An object of [TDIDriver-class],}

\item{...}{Other arguments.}
}
\description{
Provide API connection
Retrieves information on objects of classes inheriting from [TDIObject-class].
}
\seealso{
Other TDIDriver generics: 
\code{\link{TDIDriver-class}},
\code{\link{getInfo}()},
\code{\link{getSymbols}()}
}
\concept{TDIDriver generics}
