% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIBase.R
\name{baseTDI}
\alias{baseTDI}
\title{TDI Base (R6 class constructor)}
\description{
Generic and abstract R6 class within the TDI class structure.
}
\seealso{
Inspired by \href{https://dbi.r-dbi.org/}{DBI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{baseTDI$new()}}
\item \href{#method-fields}{\code{baseTDI$fields()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Abstract class initialization generates an error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baseTDI$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fields"></a>}}
\if{latex}{\out{\hypertarget{method-fields}{}}}
\subsection{Method \code{fields()}}{
Return the properties of this object as a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baseTDI$fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of object properties.
}
}
}
