% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdataAPI.R
\name{bdataAPI}
\alias{bdataAPI}
\title{Borsdata API (R6 constructor class)}
\description{
Borsdata API class inherits `TDIConnection` to implement the API requests.
}
\seealso{
API documentation on \strong{\href{https://github.com/Borsdata-Sweden/API}{Börsdata API}}.
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIConnection]{TDI::TDIConnection}} -> \code{bdataAPI}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{countries}}{Data frame with list of countries.}

\item{\code{sectors}}{Data frame with list of sectors.}

\item{\code{branches}}{Data frame with list of branches.}

\item{\code{markets}}{Data frame with list of markets.}

\item{\code{instruments}}{Data frame with list of instruments.}

\item{\code{prices}}{Latest instrument prices.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{bdataAPI$new()}}
\item \href{#method-getInstrument}{\code{bdataAPI$getInstrument()}}
\item \href{#method-getChart}{\code{bdataAPI$getChart()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIConnection" data-id="getCashFlow">}\href{../../TDI/html/TDIConnection.html#method-getCashFlow}{\code{TDI::TDIConnection$getCashFlow()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
!!!!!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bdataAPI$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{see \code{\link{TDIConnection}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `bdataAPI`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInstrument"></a>}}
\if{latex}{\out{\hypertarget{method-getInstrument}{}}}
\subsection{Method \code{getInstrument()}}{
Retrieve instrument details for the symbol.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bdataAPI$getInstrument(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{see \code{\link{TDIConnection}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `Instrument` with details.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getChart"></a>}}
\if{latex}{\out{\hypertarget{method-getChart}{}}}
\subsection{Method \code{getChart()}}{
Retrieve historical prices for the symbol.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bdataAPI$getChart(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{see \code{\link{TDIConnection}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `Instrument` with historical prices.
}
}
}
