% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdataInstrument.R
\name{bdataInstrument}
\alias{bdataInstrument}
\title{Börsdata API instrument data mapper}
\description{
Börsdata API implementation to create an instrument.
This class is a simple data mapper.
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIResult]{TDI::TDIResult}} -> \code{\link[TDI:TDIInstrument]{TDI::TDIInstrument}} -> \code{bdataInstrument}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{bdataInstrument$new()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIResult" data-id="fields">}\href{../../TDI/html/TDIResult.html#method-fields}{\code{TDI::TDIResult$fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIResult" data-id="isValid">}\href{../../TDI/html/TDIResult.html#method-isValid}{\code{TDI::TDIResult$isValid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIResult" data-id="validate">}\href{../../TDI/html/TDIResult.html#method-validate}{\code{TDI::TDIResult$validate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIInstrument" data-id="TechnicalAnalysis">}\href{../../TDI/html/TDIInstrument.html#method-TechnicalAnalysis}{\code{TDI::TDIInstrument$TechnicalAnalysis()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIInstrument" data-id="getReturn">}\href{../../TDI/html/TDIInstrument.html#method-getReturn}{\code{TDI::TDIInstrument$getReturn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIInstrument" data-id="getSession">}\href{../../TDI/html/TDIInstrument.html#method-getSession}{\code{TDI::TDIInstrument$getSession()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIInstrument" data-id="getWealthIndex">}\href{../../TDI/html/TDIInstrument.html#method-getWealthIndex}{\code{TDI::TDIInstrument$getWealthIndex()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIInstrument" data-id="setQuote">}\href{../../TDI/html/TDIInstrument.html#method-setQuote}{\code{TDI::TDIInstrument$setQuote()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIInstrument" data-id="setSeries">}\href{../../TDI/html/TDIInstrument.html#method-setSeries}{\code{TDI::TDIInstrument$setSeries()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instrument object initialization for Börsdata API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bdataInstrument$new(source, symbol, name, exchange, sector, industry, country)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{API source for the data.}

\item{\code{symbol}}{Symbol as instrument identification.}

\item{\code{name}}{Name for the instrument.}

\item{\code{exchange}}{Market for the instrument.}

\item{\code{sector}}{Main sector of the business.}

\item{\code{industry}}{Main industry of the business.}

\item{\code{country}}{Country of registration.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `Instrument`.
}
}
}
