% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsOHLCV.R
\name{Cl}
\alias{Cl}
\title{Get Close price(s).}
\usage{
Cl(x)
}
\arguments{
\item{x}{Matrix or time-series data with multiple columns.}
}
\value{
Close price(s).
}
\description{
Retrieve the Close price(s) from the data series.
}
