% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xts.R
\name{colsbind.xts}
\alias{colsbind.xts}
\title{Helper binding column(s) to time-series.}
\usage{
colsbind.xts(x, y)
}
\arguments{
\item{x}{Original xts time-series.}

\item{y}{Xts time-series column(s) to add.}
}
\value{
Xts time-series with selected column(s).
}
\description{
Helper function to bind `Y` column(s) to `X` time-series data.
Be aware that **duplicate columns in `X` get dropped!**
}
