% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xts.R
\name{columns.xts}
\alias{columns.xts}
\title{Helper to select column(s) from time-series.}
\usage{
columns.xts(x, cols)
}
\arguments{
\item{x}{Xts time-series with column(s).}

\item{cols}{List or vector with column(s).}
}
\value{
Time-series with selected column(s).
}
\description{
Helper function to select column(s) from time-series data.
}
