% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIDriver.R
\name{driver}
\alias{driver}
\title{Create driver for the source API}
\usage{
driver(source)
}
\arguments{
\item{source}{Source code to identify the API.}
}
\description{
Generic helper method to create a driver object to access the source API.
}
