% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredDriver.R
\name{fred}
\alias{fred}
\title{FRED API driver (R6 class constructor).}
\description{
Driver class to implement FRED API access, simply called `fred`.
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIDriver]{TDI::TDIDriver}} -> \code{fred}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connect}{\code{fred$connect()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIDriver" data-id="initialize">}\href{../../TDI/html/TDIDriver.html#method-initialize}{\code{TDI::TDIDriver$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect"></a>}}
\subsection{Method \code{connect()}}{
Setup API connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fred$connect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object of type `TDIConnection`.
}
}
}
