% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredAPI.R
\docType{class}
\name{fredAPI-class}
\alias{fredAPI-class}
\alias{initialize,fredAPI-method}
\alias{request,fredAPI-method}
\alias{getSymbols,fredAPI-method}
\title{FRED economic date API}
\usage{
\S4method{initialize}{fredAPI}(.Object, ...)

\S4method{request}{fredAPI}(obj, path, query)

\S4method{getSymbols}{fredAPI}(obj, symbol, range, from, to, interval)
}
\description{
FRED economic data API class inheriting from [TDIConnection-class].
This class implements the FRED API requests.
}
\seealso{
API documentation on \strong{\href{https://research.stlouisfed.org/docs/api/}{St. Louis Fed web services website}}.
}
\keyword{internal}
