% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredDriver.R
\docType{class}
\name{fred-class}
\alias{fred-class}
\alias{apiConnect,fred-method}
\title{FRED API driver class and methods.}
\usage{
\S4method{apiConnect}{fred}(obj, ...)
}
\description{
Driver class to implement FRED API access, simply called `fred`.
}
\keyword{internal}
