% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIObject.R
\name{getInfo}
\alias{getInfo}
\title{Get metadata
Retrieves information on objects of classes inheriting from [TDIObject-class].}
\usage{
getInfo(obj, ...)
}
\arguments{
\item{obj}{An object of [TDIObject-class],}

\item{...}{Other arguments.}
}
\description{
Get metadata
Retrieves information on objects of classes inheriting from [TDIObject-class].
}
\seealso{
Other TDIDriver generics: 
\code{\link{TDIDriver-class}},
\code{\link{apiConnect}()}
}
\concept{TDIDriver generics}
