% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument.R
\name{getReturn}
\alias{getReturn}
\title{Get instrument return}
\usage{
getReturn(obj)
}
\arguments{
\item{obj}{An object of [TDIResult-class].}
}
\value{
Instrument return series.
}
\description{
Get instrument return
}
