% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIResult.R
\name{getSeries}
\alias{getSeries}
\title{Get slot series}
\usage{
getSeries(obj)
}
\arguments{
\item{obj}{An object of [TDIResult-class].}
}
\value{
Object series slot.
}
\description{
Get slot series
}
\seealso{
Other TDIResult generics: 
\code{\link{TDIResult-class}},
\code{\link{addSeries}()},
\code{\link{getSymbol}()},
\code{\link{setSeries}()}
}
\concept{TDIResult generics}
