% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument.R
\name{getSession}
\alias{getSession}
\title{Get trading session}
\usage{
getSession(obj, ...)
}
\arguments{
\item{obj}{An object of [Instrument-class].}

\item{date}{Input date-time.}

\item{n}{Optional lag number of sessions.}
}
\value{
Session for date plus n.
}
\description{
Get trading session
}
\concept{Instrument generics}
