% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\name{getSymbol}
\alias{getSymbol}
\title{Retrieve symbols historical prices
Retrieves historical prices for the symbols.}
\usage{
getSymbol(
  obj,
  symbol,
  range = NULL,
  from = NULL,
  to = NULL,
  interval = NULL,
  ...
)
}
\arguments{
\item{obj}{An object of [TDIConnection-class].}

\item{symbol}{Symbol to identify the instrument.}

\item{range}{Optional period range.}

\item{from}{Optional start date of period.}

\item{to}{Optional end date of period.}

\item{interval}{Optional interval to retrieve quotes.}

\item{...}{Other arguments.}
}
\description{
Retrieve symbols historical prices
Retrieves historical prices for the symbols.
}
\seealso{
Other TDIConnection generics: 
\code{\link{TDIConnection-class}},
\code{\link{request}()}
}
\concept{TDIConnection generics}
