% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\name{getSymbols}
\alias{getSymbols}
\title{Retrieve symbols historical prices
Retrieves historical prices for the symbols.}
\usage{
getSymbols(obj, ...)
}
\arguments{
\item{obj}{An object of [TDIConnection-class],}

\item{...}{Other arguments.}
}
\description{
Retrieve symbols historical prices
Retrieves historical prices for the symbols.
}
\seealso{
Other TDIDriver generics: 
\code{\link{TDIDriver-class}},
\code{\link{apiConnect}()},
\code{\link{getInfo}()}
}
\concept{TDIDriver generics}
