% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument.R
\docType{methods}
\name{getWealthIndex}
\alias{getWealthIndex}
\alias{getWealthIndex,Instrument-method}
\title{Get instrument wealth index}
\usage{
getWealthIndex(obj)

\S4method{getWealthIndex}{Instrument}(obj)
}
\arguments{
\item{obj}{An object of class `Instrument`.}
}
\value{
Instrument wealth index series.
}
\description{
Return the object wealth index calculated from the returns.
}
