% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.OHLCV.R
\name{has.Hi}
\alias{has.Hi}
\title{Check for High price(s).}
\usage{
has.Hi(x, which = FALSE)
}
\arguments{
\item{x}{Matrix or time-series data with multiple columns.}

\item{which}{Boolean to ask for the column number.}
}
\value{
Boolean or column number.
}
\description{
Check if High price(s) exists in the data series.
}
