% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsOHLCV.R
\name{has.Lo}
\alias{has.Lo}
\title{Check for Low price(s).}
\usage{
has.Lo(x, which = FALSE)
}
\arguments{
\item{x}{Matrix or time-series data with multiple columns.}

\item{which}{Boolean to ask for the column number.}
}
\value{
Boolean or column number.
}
\description{
Check if Low price(s) exists in the data series.
}
