% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.OHLCV.R
\name{has.OHLC}
\alias{has.OHLC}
\title{Check for OHLC price(s).}
\usage{
has.OHLC(x, which = FALSE, all = TRUE)
}
\arguments{
\item{x}{Matrix or time-series data with multiple columns.}

\item{which}{Boolean to ask for the column number.}

\item{all}{Boolean to ask all at once.}
}
\value{
Boolean or column number(s).
}
\description{
Check if OHLC exists in the data series.
}
