% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.OHLCV.R
\name{has.Vo}
\alias{has.Vo}
\title{Check for trading Volume(s).}
\usage{
has.Vo(x, which = FALSE)
}
\arguments{
\item{x}{Matrix or time-series data with multiple columns.}

\item{which}{Boolean to ask for the column number.}
}
\value{
Boolean or column number.
}
\description{
Check if trading Volume(s) exists in the data series.
}
