% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.OHLCV.R
\name{Hi}
\alias{Hi}
\title{Get High price(s).}
\usage{
Hi(x)
}
\arguments{
\item{x}{Matrix or time-series data with multiple columns.}
}
\value{
High price(s).
}
\description{
Retrieve the High price(s) from the data series.
}
