% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexDriver.R
\name{iex}
\alias{iex}
\title{IEX Cloud API driver (R6 class constructor).}
\description{
Driver class to implement IEX Cloud API access, simply called `iex`.
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIDriver]{TDI::TDIDriver}} -> \code{iex}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connect}{\code{iex$connect()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="baseTDI" data-id="fields">}\href{../../TDI/html/baseTDI.html#method-fields}{\code{TDI::baseTDI$fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIDriver" data-id="initialize">}\href{../../TDI/html/TDIDriver.html#method-initialize}{\code{TDI::TDIDriver$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect"></a>}}
\if{latex}{\out{\hypertarget{method-connect}{}}}
\subsection{Method \code{connect()}}{
Setup API connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iex$connect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object of type `TDIConnection`.
}
}
}
