% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexAPI.R
\docType{class}
\name{iexAPI-class}
\alias{iexAPI-class}
\alias{request,iexAPI-method}
\alias{getSymbol,iexAPI-method}
\title{IEX Cloud API.}
\usage{
\S4method{initialize}{iexAPI}(.Object, ...)

\S4method{request}{iexAPI}(obj, path, query = NULL)

\S4method{getSymbol}{iexAPI}(obj, symbol, range, from, to, interval = obj@.conn_args$interval)
}
\description{
IEX Cloud API class inheriting from [TDIConnection-class].
This class implements the IEX Cloud API requests.

Initialize object(s) of class `iexAPI`.
}
\details{
IEX Cloud API interface functions.

TODO: other REST API's
/ref-data/region/nl/symbols?token=
/ref-data/exchange/nas/symbols?token=
/ref-data/exchanges?token=
}
\section{Slots}{

\describe{
\item{\code{.handle}}{Curl connection handle.}
}}

\seealso{
https://iexcloud.io/docs/api/#charts
}
\keyword{internal}
