% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexAPI.R
\name{iexAPI}
\alias{iexAPI}
\title{IEX Cloud API (R6 constructor class)}
\description{
IEX Cloud API class inherits `TDIConnection` to implement API requests.
}
\seealso{
API documentation on \strong{\href{https://iexcloud.io/docs/api}{IEX Cloud API website}}.
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIConnection]{TDI::TDIConnection}} -> \code{iexAPI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-getChart}{\code{iexAPI$getChart()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIConnection" data-id="getCashFlow">}\href{../../TDI/html/TDIConnection.html#method-getCashFlow}{\code{TDI::TDIConnection$getCashFlow()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIConnection" data-id="getInstrument">}\href{../../TDI/html/TDIConnection.html#method-getInstrument}{\code{TDI::TDIConnection$getInstrument()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIConnection" data-id="initialize">}\href{../../TDI/html/TDIConnection.html#method-initialize}{\code{TDI::TDIConnection$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getChart"></a>}}
\if{latex}{\out{\hypertarget{method-getChart}{}}}
\subsection{Method \code{getChart()}}{
Retrieve historical prices for the symbol.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iexAPI$getChart(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{see \code{\link{TDIConnection}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `Instrument` with historical prices.
}
}
}
