% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredDriver.R, R/iexDriver.R
\docType{class}
\name{IEXDriver}
\alias{IEXDriver}
\alias{iex-class}
\alias{apiConnect,iex-method}
\title{FRED economic data API driver}
\usage{
IEXDriver()

IEXDriver()

\S4method{apiConnect}{iex}(obj, ...)
}
\description{
Helper function to establish a connection with FRED API.

Helper function to establish a connection with IEX Cloud API.

Driver class to implement IEX Cloud API access, simply called `iex`.
}
\keyword{internal}
