% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexInstrument.R
\name{iexInstrument}
\alias{iexInstrument}
\title{IEX instrument data mapper}
\description{
IEX Cloud API implementation to create an instrument.
This class is a simple data mapper.
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIResult]{TDI::TDIResult}} -> \code{\link[TDI:TDIInstrument]{TDI::TDIInstrument}} -> \code{iexInstrument}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{iexInstrument$new()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIResult" data-id="fields">}\href{../../TDI/html/TDIResult.html#method-fields}{\code{TDI::TDIResult$fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIInstrument" data-id="getReturn">}\href{../../TDI/html/TDIInstrument.html#method-getReturn}{\code{TDI::TDIInstrument$getReturn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIInstrument" data-id="getSession">}\href{../../TDI/html/TDIInstrument.html#method-getSession}{\code{TDI::TDIInstrument$getSession()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIInstrument" data-id="getWealthIndex">}\href{../../TDI/html/TDIInstrument.html#method-getWealthIndex}{\code{TDI::TDIInstrument$getWealthIndex()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIInstrument" data-id="setSeries">}\href{../../TDI/html/TDIInstrument.html#method-setSeries}{\code{TDI::TDIInstrument$setSeries()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Instrument object initialization for IEX Cloud API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iexInstrument$new(source, symbol, json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{API source for the data.}

\item{\code{symbol}}{Symbol as instrument identification.}

\item{\code{json}}{Json string returned by the API.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `Instrument`.
}
}
}
