% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indicator.R
\name{Indicator}
\alias{Indicator}
\title{Indicator (R6 class constructor)}
\description{
Indicator is a R6 class for economic data and other indicator(s).
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIResult]{TDI::TDIResult}} -> \code{Indicator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{Indicator type classification.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Indicator$new()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIResult" data-id="setSeries">}\href{../../TDI/html/TDIResult.html#method-setSeries}{\code{TDI::TDIResult$setSeries()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Constructor for object(s) of class `Indicator`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Indicator$new(source, symbol, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{API source for the data.}

\item{\code{symbol}}{Unique identification symbol.}

\item{\code{type}}{Indicator type classification.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `Indicator`.
}
}
}
