% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIInstrument.R
\name{Instrument}
\alias{Instrument}
\title{Instrument constructor}
\usage{
Instrument(symbol, source)
}
\arguments{
\item{symbol}{Symbol to identify the instrument.}

\item{sources}{Sources for instrument data.}
}
\value{
Object of class TDIInstrument.
}
\description{
Helper constructor for object(s) of class TDIInstrument.
}
