% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIBase.R
\name{is.baseTDI}
\alias{is.baseTDI}
\title{Check for base class of TDI}
\usage{
is.baseTDI(x = NULL)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Boolean result or alternative the class name.
}
\description{
Helper to check the object class.
If the input `x` is empty it returns the class name.
}
