% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indicator.R
\name{is.Indicator}
\alias{is.Indicator}
\title{Check if the object is of class Indicator}
\usage{
is.Indicator(x = NULL)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Boolean result or alternative the class name.
}
\description{
Helper to check the object class.
If the input `x` is empty it returns the class name.
}
