% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIInstrument.R
\name{is.Instrument}
\alias{is.Instrument}
\title{Check if the object is of class Instrument}
\usage{
is.Instrument(x = NULL)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Boolean result or alternative the class name.
}
\description{
Helper to check the object class.
If the input `x` is empty it returns the class name.
}
