% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAIndicator.R
\name{is.TAIndicator}
\alias{is.TAIndicator}
\title{Check for TA Indicator class}
\usage{
is.TAIndicator(x = NULL)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Boolean result or alternative the class name.
}
\description{
Helper to check the object class.
If the input `x` is empty it returns the class name.
}
