% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIDriver.R
\name{is.TDIDriver}
\alias{is.TDIDriver}
\title{Check for class TDIDriver}
\usage{
is.TDIDriver(x = NULL)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Boolean result or alternative the class name.
}
\description{
Helper to check the object class.
If the input `x` is empty it returns the class name.
}
