% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.OHLCV.R
\name{Lo}
\alias{Lo}
\title{Get Low price(s).}
\usage{
Lo(x)
}
\arguments{
\item{x}{Matrix or time-series data with multiple columns.}
}
\value{
Low price(s).
}
\description{
Retrieve the Low price(s) from the data series.
}
