% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match.String}
\alias{match.String}
\title{Helper to match string(s).}
\usage{
match.String(pattern = null, x, value = FALSE)
}
\arguments{
\item{pattern}{String or vector with pattern(s) to match.}

\item{x}{String or vector to search.}

\item{value}{Boolean to retrieve the value.}
}
\value{
Vector with column numbers or names.
}
\description{
Helper function to return a vector with string(s) matching the pattern(s).
The function aims to match each pattern individually (in a loop),
and returns an error if one of them has no match.
}
\details{
The function first attempts an exact match.
When there are no exact matches it attempts to match a regular expression.
The regular expression aims to match at the end of the input `x`.

In case of multiple matches on the first result is returned.
Depending on the `value` a value or index is returned.
}
\examples{
match.String(c("A", "b"), c("A", "B", "C")) returns [1] 1 2
}
