% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.OHLCV.R
\name{OHLC}
\alias{OHLC}
\title{Get OHLC price(s).}
\usage{
OHLC(x)
}
\arguments{
\item{x}{Matrix or time-series data with multiple columns.}
}
\value{
OHLC price(s).
}
\description{
Retrieve OHLC price(s) from the data series.
}
