% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xts.R
\name{OHLC.xts}
\alias{OHLC.xts}
\title{Helper extracting OHLC data.}
\usage{
OHLC.xts(x)
}
\arguments{
\item{x}{Time-series with OHCL.}
}
\value{
Time-series with only OHCL.
}
\description{
Helper function to extract the OHCL data from time-series.
}
