% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsOHLCV.R
\name{Op}
\alias{Op}
\title{Get Open price(s).}
\usage{
Op(x)
}
\arguments{
\item{x}{Matrix or time-series data with multiple columns.}
}
\value{
Open price(s).
}
\description{
Retrieve the Open price(s) from the data series.
}
