% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\docType{methods}
\name{request.JSON}
\alias{request.JSON}
\alias{request.JSON,TDIConnection-method}
\title{Execute API request with JSON}
\usage{
request.JSON(obj, url, query)

\S4method{request.JSON}{TDIConnection}(obj, url, query = NULL)
}
\arguments{
\item{obj}{An object of class `TDIConnection`.}

\item{url}{Final URL for the endpoint.}

\item{query}{URL query arguments.}
}
\description{
Generic helper method to execute JSON requests.
}
