% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\docType{methods}
\name{request}
\alias{request}
\title{Execute API request}
\usage{
request(obj, path, query)
}
\arguments{
\item{obj}{An object of class `TDIConnection`.}

\item{path}{Relative path for the endpoint.}

\item{query}{URL query arguments.}
}
\description{
Call the API endpoint with the relative URL path and query parameters.
}
\seealso{
Other TDIConnection generics: 
\code{\link{TDIConnection-class}},
\code{\link{getChart}()},
\code{\link{validInterval}()},
\code{\link{validRange}()}
}
\concept{TDIConnection generics}
