% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIResult.R
\name{setSeries}
\alias{setSeries}
\title{Set slot series.}
\usage{
setSeries(obj, df)
}
\arguments{
\item{obj}{An object of [TDIResult-class].}

\item{df}{Data frame to be saved.}
}
\value{
Object of class `TDIResult`.
}
\description{
Set slot series.
}
\seealso{
Other TDIResult generics: 
\code{\link{TDIResult-class}}
}
\concept{TDIResult generics}
