% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAIndicator.R
\name{taFactory}
\alias{taFactory}
\title{Technical analysis indicator factory}
\usage{
taFactory(ind)
}
\arguments{
\item{ind}{Indicator definition.}
}
\value{
An object of class `TAIndicator`.
}
\description{
Constructor to instantiate object(s) of class `TAIndicator`.
It maps the indicator definition to the class properties.
}
\examples{
# Call the factory with SMA setup.
json <- '{
  "label": "SMA",
  "method": "TTR::SMA",
  "columns": [{"x": ["Close"]}]
}'
ind <- jsonlite::fromJSON(json)
ta <- taFactory(ind)

# Call the factory with ATR setup.
json <- '{
  "label": "ATR",
  "method": "TTR::ATR",
  "columns": [{"HLC": ["High", "Low", "Close"]}],
  "args": [{"n": 14, "maType": "EMA"}]
}'
ind <- jsonlite::fromJSON(json)
ta <- taFactory(ind)

}
