% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAIndicator.R
\name{TAIndicator}
\alias{TAIndicator}
\title{Technical analysis indicator (R6 class constructor)}
\description{
An R6 class for technical analysis indicator(s).
Its purpose is to facilitate calculations on time-series of data.
The fields specify the `func` and parameters for the calculation(s).
The indicator(s) can be constructed with the function `taFactory()`.
}
\section{Super class}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{TAIndicator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label}}{Unique label for the indicator.}

\item{\code{method}}{Indicator calculation method.}

\item{\code{columns}}{List with input data column(s).}

\item{\code{args}}{Arguments for the calculation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TAIndicator$new()}}
\item \href{#method-calculate}{\code{TAIndicator$calculate()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Object constructor for class `taIndicator`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TAIndicator$new(label, method, columns, args = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{Label for the indicator.}

\item{\code{method}}{Method to calculate the indicator.}

\item{\code{columns}}{List of input data column(s).}

\item{\code{args}}{List of arguments for the indicator function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `taIndicator`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate"></a>}}
\if{latex}{\out{\hypertarget{method-calculate}{}}}
\subsection{Method \code{calculate()}}{
Calculate the indicator for the time series.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TAIndicator$calculate(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Time-series object with input data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
xts Time-serie with the indicator(s).
}
}
}
