% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\docType{class}
\name{TDIConnection-class}
\alias{TDIConnection-class}
\title{TDIConnection class}
\description{
Wraps objects of the [TDIDriver-class] class with connection options.
The purpose is to store both the driver and the connection options.

API access can be established through [apiConnect()], 
passing only the object with optional additional arguments.
}
\examples{
# TO-DO
}
\seealso{
Other TDI classes: 
\code{\link{TDIDriver-class}},
\code{\link{TDIObject-class}}
}
\concept{TDI classes}
\concept{TDIConnection generics}
