% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnector.R
\docType{class}
\name{TDIConnector-class}
\alias{TDIConnector-class}
\alias{TDIConnector}
\title{TDIConnector}
\format{An object of class \code{TDIConnector} of length 1.}
\usage{
TDIConnector
}
\description{
The `TDIConnector` manages `TDIConnection`(s).
It's purpose is to establish connection(s) to acccess API(s) for trading data.
Each API has its own driver that is identified by the `source` (class).

The `connect()` method returns a connection to be used for API access,
which is an object of class `TDIConnection`. 
Based upon the `source` (driver class) a driver object of class [TDIDriver-class] is loaded. 
The driver initialized a connection object of class [TDIConnection-class].

The class implements the **singleton pattern** to re-use connections. 
Note that the class is immediately instantiated.
}
\section{Fields}{

\describe{
\item{\code{.connections}}{List with connection pool.}
}}

\examples{
# TDI library creates an singleton object of class `TDIConnector`.
# It's `connect()` method is to be used to establish a connection,
# that faciliates the API request.
con <- TDI::TDIConnector$connect("yahoo")
getSymbols(con, "MSFT")

}
\seealso{
Other TDI classes: 
\code{\link{Instrument-class}},
\code{\link{TDIConnection-class}},
\code{\link{TDIDriver-class}},
\code{\link{TDIIndicator-class}},
\code{\link{TDIObject-class}},
\code{\link{TDIResult-class}}
}
\concept{TDI classes}
\keyword{datasets}
