% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnector.R
\docType{data}
\name{TDIConnector}
\alias{TDIConnector}
\title{TDIConnector}
\format{
An object of class \code{TDIConnector} (inherits from \code{R6}) of length 3.
}
\usage{
TDIConnector(source = "yahoo")
}
\arguments{
\item{source}{API driver class name.}
}
\value{
An object of class `TDIConnection`.
}
\description{
TDIConnector is an instantiation of class `TDIConnector` to manage connection(s).
It's purpose is to establish connection(s) to acccess API(s) for qualitative data.

Setup and provide an API connection.
}
\details{
The `TDIConnector` connects with an API through its driver identified by `source` (class).
It's `connect()` method establishes the API access using the `TDIDriver`,
and returns that connection as an object of class `TDIConnection`. 

The class implements the **singleton pattern** to re-use connections. 
Note that the class is immediately instantiated.
}
\examples{
# TDI library creates an singleton object of class `TDIConnector`.
# It's `connect()` method is to be used to establish a connection,
# that facilitates the API request.
con <- TDI::TDIConnector$connect(source = "yahoo")
msft <- con$getChart(symbol = "MSFT")
}
\keyword{datasets}
