% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIDriver.R
\docType{class}
\name{TDIDriver-class}
\alias{TDIDriver-class}
\alias{getInfo,TDIDriver-method}
\title{TDIDriver class}
\usage{
\S4method{getInfo}{TDIDriver}(obj)
}
\arguments{
\item{obj}{An object of class `TDIDriver`.}
}
\description{
Base class for all TDI drivers (e.g., Yahoo, IEX, etc).
The virtual class `TDIDriver` is the generic class for API specific driver(s).
Each API has a driver sub-class of `TDIDriver` class to initiate the API.
}
\seealso{
Other TDI classes: 
\code{\link{Instrument-class}},
\code{\link{TDIConnection-class}},
\code{\link{TDIConnector-class}},
\code{\link{TDIIndicator-class}},
\code{\link{TDIObject-class}},
\code{\link{TDIResult-class}}

Other TDIDriver generics: 
\code{\link{apiConnect}()},
\code{\link{getInfo}()}
}
\concept{TDI classes}
\concept{TDIDriver generics}
