% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIIndicator.R
\docType{class}
\name{TDIIndicator-class}
\alias{TDIIndicator-class}
\title{TDIIndicator class}
\description{
Indicator class is a [TDIResult-class] for economic data and other indicator(s).
}
\section{Slots}{

\describe{
\item{\code{.type}}{Type classification.}
}}

\seealso{
Other TDI classes: 
\code{\link{TDIConnection-class}},
\code{\link{TDIConnector-class}},
\code{\link{TDIDriver-class}},
\code{\link{TDIInstrument-class}},
\code{\link{TDIObject-class}},
\code{\link{TDIResult-class}}
}
\concept{TDI classes}
