% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIInstrument.R
\docType{class}
\name{TDIInstrument-class}
\alias{TDIInstrument-class}
\title{TDIInstrument class}
\description{
Instrument class is a [TDIResult-class] for all types of (tradeable) market instruments.
}
\section{Slots}{

\describe{
\item{\code{.type}}{Type classification.}
}}

\seealso{
Other TDI classes: 
\code{\link{TDIConnection-class}},
\code{\link{TDIConnector-class}},
\code{\link{TDIDriver-class}},
\code{\link{TDIIndicator-class}},
\code{\link{TDIObject-class}},
\code{\link{TDIResult-class}}
}
\concept{TDI classes}
