% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIInstrument.R
\name{TDIInstrument}
\alias{TDIInstrument}
\title{Instrument (R6 class constructor)}
\description{
Instrument is an abstract class for (tradeable) market instrument(s).
The class provides generic methods and defines the instrument data structure.
}
\details{
The instrument class and its generic methods are to be used by calling packages.
The class is abstract and should be implemented for each API.
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIResult]{TDI::TDIResult}} -> \code{TDIInstrument}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{source}}{Identifier for the data source.}

\item{\code{symbol}}{Unique identification symbol.}

\item{\code{type}}{Instrument type classification.}

\item{\code{name}}{Name for the instrument.}

\item{\code{exchange}}{Exchange for the instrument.}

\item{\code{sector}}{Business sector.}

\item{\code{industry}}{Business industry.}

\item{\code{country}}{Country of registration.}

\item{\code{keyData}}{Key data object.}

\item{\code{annualCashFlow}}{Annual cash flow data.}

\item{\code{quarterlyCashFlow}}{Quarterly cash flow data.}

\item{\code{series}}{Xts time series with (historical) data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-setSeries}{\code{TDIInstrument$setSeries()}}
\item \href{#method-getSession}{\code{TDIInstrument$getSession()}}
\item \href{#method-getReturn}{\code{TDIInstrument$getReturn()}}
\item \href{#method-getWealthIndex}{\code{TDIInstrument$getWealthIndex()}}
\item \href{#method-TechnicalAnalysis}{\code{TDIInstrument$TechnicalAnalysis()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="baseTDI" data-id="initialize">}\href{../../TDI/html/baseTDI.html#method-initialize}{\code{TDI::baseTDI$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIResult" data-id="fields">}\href{../../TDI/html/TDIResult.html#method-fields}{\code{TDI::TDIResult$fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIResult" data-id="isValid">}\href{../../TDI/html/TDIResult.html#method-isValid}{\code{TDI::TDIResult$isValid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIResult" data-id="validate">}\href{../../TDI/html/TDIResult.html#method-validate}{\code{TDI::TDIResult$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSeries"></a>}}
\if{latex}{\out{\hypertarget{method-setSeries}{}}}
\subsection{Method \code{setSeries()}}{
Set series or add additional column(s) to the existing data.
In case the input is a data-frame it is converted to `xts`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIInstrument$setSeries(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Xts time series with (historical) data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object itself.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSession"></a>}}
\if{latex}{\out{\hypertarget{method-getSession}{}}}
\subsection{Method \code{getSession()}}{
Retrieve the trading session data for the input date plus `n` periods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIInstrument$getSession(date, n = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{Input date-time.}

\item{\code{n}}{Optional lag number of sessions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Requested session data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getReturn"></a>}}
\if{latex}{\out{\hypertarget{method-getReturn}{}}}
\subsection{Method \code{getReturn()}}{
Return the *returns* calculated from the `Close` price(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIInstrument$getReturn()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Instrument returns time series.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getWealthIndex"></a>}}
\if{latex}{\out{\hypertarget{method-getWealthIndex}{}}}
\subsection{Method \code{getWealthIndex()}}{
Return the *wealth index* calculated from the returns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIInstrument$getWealthIndex()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Instrument returns time series.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TechnicalAnalysis"></a>}}
\if{latex}{\out{\hypertarget{method-TechnicalAnalysis}{}}}
\subsection{Method \code{TechnicalAnalysis()}}{
Execute the specified technical analysis calculations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIInstrument$TechnicalAnalysis(ta = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ta}}{Technical analysis indicator object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object itself.
}
}
}
