% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIKeyData.R
\name{TDIKeyData}
\alias{TDIKeyData}
\title{TDIKeyData (R6 class constructor)}
\description{
Generic abstract base class for TDI result(s) key data.
The class is a data structure for instrument key data,
and should be implemented in the API implementation.
Using R6 classes for data structures enforce standardization.
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIResult]{TDI::TDIResult}} -> \code{TDIKeyData}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sharesOutstanding}}{Total number of shares.}

\item{\code{currency}}{Denomination currency.}

\item{\code{price}}{Current price.}

\item{\code{marketCap}}{Shares total market capitalization.}

\item{\code{ebitda}}{EBITDA.}

\item{\code{trailingEPS}}{Trailing earnings per share.}

\item{\code{forwardEPS}}{Forward earnings per share.}

\item{\code{trailingPE}}{Trailing price/earnings.}

\item{\code{forwardPE}}{Forward price/earnings.}

\item{\code{PEG}}{Price/earnings to growth ratio.}
}
\if{html}{\out{</div>}}
}
