% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIObject.R
\docType{class}
\name{TDIObject-class}
\alias{TDIObject-class}
\title{TDIObject class}
\description{
Base class for all TDI classes (e.g., drivers, etc). This
is a virtual Class: No objects may be created from it.

More generally, the TDI defines a small set of classes and generics to
facilitate a common interface for trading data.
}
\seealso{
Inspired by \href{https://dbi.r-dbi.org/}{DBI}

Other TDI classes: 
\code{\link{Instrument-class}},
\code{\link{TDIConnection-class}},
\code{\link{TDIConnector-class}},
\code{\link{TDIDriver-class}},
\code{\link{TDIIndicator-class}},
\code{\link{TDIResult-class}}
}
\concept{TDI classes}
