% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIResult.R
\name{TDIResult}
\alias{TDIResult}
\title{TDIResult (R6 class constructor)}
\description{
Generic base class for TDI result(s) sub-classes (e.g. Instrument, Index).
}
\section{Super class}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{TDIResult}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sources}}{Identifier for data source.}

\item{\code{symbol}}{Unique identification symbol.}

\item{\code{series}}{Xts time-series with (historical) data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-setSeries}{\code{TDIResult$setSeries()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="baseTDI" data-id="fields">}\href{../../TDI/html/baseTDI.html#method-fields}{\code{TDI::baseTDI$fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="baseTDI" data-id="initialize">}\href{../../TDI/html/baseTDI.html#method-initialize}{\code{TDI::baseTDI$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSeries"></a>}}
\if{latex}{\out{\hypertarget{method-setSeries}{}}}
\subsection{Method \code{setSeries()}}{
Set series or add additional column(s) to the existing data.
In case the input is a data-frame it is converted to `xts`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIResult$setSeries(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Xts time-series with (historical) data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `TDIResult`.
}
}
}
