% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\docType{methods}
\name{validInterval}
\alias{validInterval}
\title{Validate the API interval parameter}
\usage{
validInterval(obj, interval = NULL)
}
\arguments{
\item{obj}{An object of class `TDIConnection`.}

\item{interval}{Interval query parameter.}
}
\description{
Generic validation method for API interval parameter.
}
\seealso{
Other TDIConnection generics: 
\code{\link{TDIConnection-class}},
\code{\link{getChart}()},
\code{\link{request}()},
\code{\link{validRange}()}
}
\concept{TDIConnection generics}
