% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\name{validRange}
\alias{validRange}
\title{Validate the API range parameter
Generic validation method for API range parameter.}
\usage{
validRange(obj, range = NULL)
}
\arguments{
\item{obj}{An object of [TDIConnection-class].}

\item{range}{Range query parameter.}
}
\description{
Validate the API range parameter
Generic validation method for API range parameter.
}
\seealso{
Other TDIConnection generics: 
\code{\link{TDIConnection-class}},
\code{\link{getSymbols}()},
\code{\link{request}()},
\code{\link{validInterval}()}
}
\concept{TDIConnection generics}
