% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsOHLCV.R
\name{Vo}
\alias{Vo}
\title{Get trading Volume(s).}
\usage{
Vo(x)
}
\arguments{
\item{x}{Matrix or time-series data with multiple columns.}
}
\value{
Volume(s) traded.
}
\description{
Retrieve the trading Volume(s) from the data series.
}
