% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahooAPI.R
\docType{class}
\name{YahooFin-class}
\alias{YahooFin-class}
\alias{getSymbols,YahooFin-method}
\title{Yahoo Finance API.}
\usage{
\S4method{getSymbols}{YahooFin}(obj, symbol)
}
\description{
Yahoo Finance API class inheriting from [TDIConnection-class].
Objects of this class provide the Yahoo API methods.
}
\keyword{internal}
