% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahooCashFlow.R
\name{yahooCashFlow}
\alias{yahooCashFlow}
\title{Yahoo Key Data mapper (R6 class constructor)}
\description{
Yahoo Finance API implementation to retrieve cash flow data.
This class is a simple data mapper.
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIResult]{TDI::TDIResult}} -> \code{\link[TDI:TDICashFlow]{TDI::TDICashFlow}} -> \code{yahooCashFlow}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{yahooCashFlow$new()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIResult" data-id="fields">}\href{../../TDI/html/TDIResult.html#method-fields}{\code{TDI::TDIResult$fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIResult" data-id="isValid">}\href{../../TDI/html/TDIResult.html#method-isValid}{\code{TDI::TDIResult$isValid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIResult" data-id="validate">}\href{../../TDI/html/TDIResult.html#method-validate}{\code{TDI::TDIResult$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialization with Yahoo Finance cash flow data mapping.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{yahooCashFlow$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Subset of data returned by the API.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `YahooCashFlow`.
}
}
}
