% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahooDriver.R
\docType{class}
\name{yahoo-class}
\alias{yahoo-class}
\alias{apiConnect,yahoo-method}
\title{Yahoo Finance API driver class and methods.}
\usage{
\S4method{apiConnect}{yahoo}(obj, ...)
}
\description{
Driver class to implement Yahoo Finance API access, simply called `yahoo`.
}
\keyword{internal}
