% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahooKeyData.R
\name{yahooKeyData}
\alias{yahooKeyData}
\title{Yahoo Key Data mapper (R6 class constructor)}
\description{
Yahoo Finance API implementation to retrieve instrument key data.
This class is a simple data mapper.
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIKeyData]{TDI::TDIKeyData}} -> \code{yahooKeyData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{yahooKeyData$new()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="baseTDI" data-id="fields">}\href{../../TDI/html/baseTDI.html#method-fields}{\code{TDI::baseTDI$fields()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialization with Yahoo Finance instrument key data mapping.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{yahooKeyData$new(json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{json}}{Json string returned by the API.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `YahooKeyData`.
}
}
}
