% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIDriver.R
\docType{methods}
\name{apiConnect}
\alias{apiConnect}
\title{Provide API connection}
\usage{
apiConnect(obj, ...)
}
\arguments{
\item{obj}{An object of class `TDIDriver`.}

\item{...}{Other arguments.}
}
\description{
Retrieves information on objects of classes inheriting from [TDIObject-class].
}
\seealso{
Other TDIDriver generics: 
\code{\link{TDIDriver-class}},
\code{\link{getInfo}()}
}
\concept{TDIDriver generics}
