% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIObject.R
\docType{methods}
\name{getInfo}
\alias{getInfo}
\title{Get metadata}
\usage{
getInfo(obj, ...)
}
\arguments{
\item{obj}{An object of class `TDIObject`.}

\item{...}{Other arguments.}
}
\description{
Retrieves information on objects of classes inheriting from [TDIObject-class].
}
\seealso{
Other TDIDriver generics: 
\code{\link{TDIDriver-class}},
\code{\link{apiConnect}()}
}
\concept{TDIDriver generics}
