% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument.R
\docType{methods}
\name{getReturn}
\alias{getReturn}
\alias{getReturn,Instrument-method}
\title{Get instrument return}
\usage{
getReturn(obj)

\S4method{getReturn}{Instrument}(obj)
}
\arguments{
\item{obj}{An object of class `Instrument`.}
}
\value{
Instrument return series.
}
\description{
Return the object *returns* calculated from the `Close` price(s).
}
