% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\name{getSeries}
\alias{getSeries}
\title{Retrieve historical data time series.}
\usage{
getSeries(
  obj,
  symbol,
  range = NULL,
  from = NULL,
  to = NULL,
  interval = NULL,
  ...
)
}
\arguments{
\item{obj}{An object of [TDIConnection-class].}

\item{symbol}{Symbol to identify the instrument.}

\item{range}{Optional period range.}

\item{from}{Optional start date of period.}

\item{to}{Optional end date of period.}

\item{interval}{Optional interval to retrieve quotes.}

\item{...}{Other arguments.}
}
\description{
Generic S4 method to retrieve historical data time series for the symbol.
Sub-classes implements the method for a specific trading data API.
}
\seealso{
Other TDIConnection generics: 
\code{\link{TDIConnection-class}},
\code{\link{request}()},
\code{\link{validInterval}()},
\code{\link{validRange}()}
}
\concept{TDIConnection generics}
