% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIResult.R
\docType{methods}
\name{getSeries}
\alias{getSeries}
\alias{getSeries,TDIResult-method}
\title{Get series for `TDIResult`}
\usage{
getSeries(obj)

\S4method{getSeries}{TDIResult}(obj)
}
\arguments{
\item{obj}{An object of class `TDIResult`.}
}
\value{
Series as `xts` time-series.
}
\description{
Return the series for input object of class `TDIResult`.
}
\seealso{
Other TDIResult generics: 
\code{\link{TDIResult-class}},
\code{\link{getSymbol}()},
\code{\link{setSeries}()}
}
\concept{TDIResult generics}
