% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument.R
\docType{methods}
\name{getSession}
\alias{getSession}
\alias{getSession,Instrument-method}
\title{Get trading session}
\usage{
getSession(obj, date, n = 0)

\S4method{getSession}{Instrument}(obj, date, n = 0)
}
\arguments{
\item{obj}{An object of class `Instrument`.}

\item{date}{Input date-time.}

\item{n}{Optional lag number of sessions.}
}
\value{
Requested session data.
}
\description{
Return the tradung session data for the input date plus `n` periods.
}
\concept{Instrument generics}
