% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIResult.R
\name{getSymbol}
\alias{getSymbol}
\title{Get symbol [TDIResult-class]}
\usage{
getSymbol(obj)
}
\arguments{
\item{obj}{An object of [TDIResult-class].}
}
\value{
Symbol as character.
}
\description{
Get symbol [TDIResult-class]
}
\seealso{
Other TDIResult generics: 
\code{\link{TDIResult-class}},
\code{\link{addSeries}()},
\code{\link{getSeries}()},
\code{\link{setSeries}()}
}
\concept{TDIResult generics}
