% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\name{getSymbol}
\alias{getSymbol}
\title{Retrieve symbols historical prices
Retrieves historical prices for the symbols.}
\usage{
getSymbol(obj, symbol, ...)
}
\arguments{
\item{obj}{An object of [TDIConnection-class].}

\item{symbol}{Symbol to identify the instrument.}

\item{...}{Other arguments.}
}
\description{
Retrieve symbols historical prices
Retrieves historical prices for the symbols.
}
\seealso{
Other TDIConnection generics: 
\code{\link{TDIConnection-class}}
}
\concept{TDIConnection generics}
