% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIResult.R
\docType{methods}
\name{getSymbol}
\alias{getSymbol}
\alias{getSymbol,TDIResult-method}
\title{Get symbol for `TDIResult`}
\usage{
getSymbol(obj)

\S4method{getSymbol}{TDIResult}(obj)
}
\arguments{
\item{obj}{An object of class `TDIResult`.}
}
\value{
Symbol as character.
}
\description{
Return the symbol for input object of class `TDIResult`.
}
\seealso{
Other TDIResult generics: 
\code{\link{TDIResult-class}},
\code{\link{getSeries}()},
\code{\link{setSeries}()}
}
\concept{TDIResult generics}
