% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument.R
\name{getWealthIndex}
\alias{getWealthIndex}
\title{Get instrument wealth index}
\usage{
getWealthIndex(obj)
}
\arguments{
\item{obj}{An object of [TDIResult-class].}
}
\value{
Instrument wealth index series.
}
\description{
Get instrument wealth index
}
