% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexAPI.R
\docType{class}
\name{iexAPI-class}
\alias{iexAPI-class}
\alias{initialize,iexAPI-method}
\alias{request,iexAPI-method}
\alias{validRange,iexAPI-method}
\alias{validInterval,iexAPI-method}
\alias{getSymbols,iexAPI-method}
\title{IEX Cloud API}
\usage{
\S4method{initialize}{iexAPI}(.Object, ...)

\S4method{request}{iexAPI}(obj, path, query)

\S4method{validRange}{iexAPI}(obj, range = NULL)

\S4method{validInterval}{iexAPI}(obj, interval = NULL)

\S4method{getSymbols}{iexAPI}(obj, symbol, range, from, to, interval)
}
\description{
IEX Cloud API class inheriting from [TDIConnection-class].
This class implements the IEX Cloud API requests.
}
\seealso{
API documentation on \strong{\href{https://iexcloud.io/docs/api}{IEX Cloud API website}}.
}
\keyword{internal}
