% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexDriver.R
\docType{class}
\name{iex-class}
\alias{iex-class}
\alias{apiConnect,iex-method}
\title{IEX Cloud API driver class and methods.}
\usage{
\S4method{apiConnect}{iex}(obj, ...)
}
\description{
Driver class to implement IEX Cloud API access, simply called `iex`.
}
\keyword{internal}
