% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexDriver.R
\docType{class}
\name{IEXDriver}
\alias{IEXDriver}
\alias{iex-class}
\alias{apiConnect,iex-method}
\title{IEX Cloud API driver}
\usage{
IEXDriver()

\S4method{apiConnect}{iex}(obj, ...)
}
\description{
Helper function to establish a connection with IEX Cloud API.

Driver class to implement IEX Cloud API access.

The driver class `iex` is named with commonly used source name.
This makes it simpler to keep references to the driver with the data.
}
\keyword{internal}
