% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument.R
\name{Instrument}
\alias{Instrument}
\title{Instrument (R6 class constructor)}
\description{
Instrument is a R6 class for all types of (tradeable) market instrument(s).
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIResult]{TDI::TDIResult}} -> \code{Instrument}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{Indicator type classification.}

\item{\code{currency}}{Nomination currency.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Instrument$new()}}
\item \href{#method-setSeries}{\code{Instrument$setSeries()}}
\item \href{#method-getSession}{\code{Instrument$getSession()}}
\item \href{#method-getReturn}{\code{Instrument$getReturn()}}
\item \href{#method-getWealthIndex}{\code{Instrument$getWealthIndex()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Constructor for object(s) of class `Indicator`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$new(source, symbol, type = NULL, currency = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{API source for the data.}

\item{\code{symbol}}{Unique identification symbol.}

\item{\code{type}}{Indicator type classification.}

\item{\code{currency}}{Nomination currency.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `Indicator`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSeries"></a>}}
\subsection{Method \code{setSeries()}}{
Set series and calculate the return from the close price.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$setSeries(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Xts time-series with (historical) data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `TDIResult`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSession"></a>}}
\subsection{Method \code{getSession()}}{
Retrieve the trading session data for the input date plus `n` periods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$getSession(date, n = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{Input date-time.}

\item{\code{n}}{Optional lag number of sessions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Requested session data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getReturn"></a>}}
\subsection{Method \code{getReturn()}}{
Return the *returns* calculated from the `Close` price(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$getReturn()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Instrument returns time-series.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getWealthIndex"></a>}}
\subsection{Method \code{getWealthIndex()}}{
Return the *wealth index* calculated from the returns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$getWealthIndex()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Instrument returns time-series.
}
}
}
