% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument.R
\name{Instrument}
\alias{Instrument}
\title{Instrument constructor}
\usage{
Instrument(symbol, source, currency = NULL, type = NULL)
}
\arguments{
\item{symbol}{Symbol to identify the instrument.}

\item{source}{Source for instrument data.}

\item{currency}{Denomination currency.}

\item{type}{Type of instrument.}
}
\value{
Object of class Instrument.
}
\description{
Helper constructor for object(s) of class Instrument.
}
