% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument.R
\name{is.Instrument}
\alias{is.Instrument}
\title{Check if the object is an instrument.}
\usage{
is.Instrument(x = NULL)
}
\arguments{
\item{x}{Object to check the class for.}
}
\value{
Boolean result for the class check or alternative the class name.
}
\description{
This helper function checks if the object is of class Instrument.
If the input `x` is empty it returns the class name.
}
