% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\name{request}
\alias{request}
\title{Execute API request
Call the API endpoint with the relative URL path and query parameters.}
\usage{
request(obj, path, query)
}
\arguments{
\item{obj}{An object of [TDIConnection-class].}

\item{path}{Relative path for the endpoint.}

\item{query}{URL query arguments.}
}
\description{
Execute API request
Call the API endpoint with the relative URL path and query parameters.
}
\seealso{
Other TDIConnection generics: 
\code{\link{TDIConnection-class}},
\code{\link{getSeries}()},
\code{\link{validInterval}()},
\code{\link{validRange}()}
}
\concept{TDIConnection generics}
