% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIResult.R, R/Instrument.R
\docType{methods}
\name{setSeries}
\alias{setSeries}
\alias{setSeries,TDIResult-method}
\alias{setSeries,Instrument-method}
\title{Set series for `TDIResult`}
\usage{
setSeries(obj, x)

\S4method{setSeries}{TDIResult}(obj, x)

\S4method{setSeries}{Instrument}(obj, x)
}
\arguments{
\item{obj}{An object of class `TDIResult`.}

\item{x}{Data for slot series.}
}
\value{
Object of class `TDIResult` with series.
}
\description{
Set series for `TDIResult`
}
\seealso{
Other TDIResult generics: 
\code{\link{TDIResult-class}},
\code{\link{addSerie}()},
\code{\link{getSeries}()},
\code{\link{getSymbol}()}
}
\concept{TDIResult generics}
