% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAIndicator.R
\name{taFactory}
\alias{taFactory}
\title{Technical analysis indicator factory}
\usage{
taFactory(label, method, columns, args = NULL)
}
\arguments{
\item{label}{Label for the indicator.}

\item{method}{Method to calculate the indicator.}

\item{columns}{Input data column(s).}

\item{args}{Input parameters for the indicator.}
}
\value{
An object of class `TAIndicator`.
}
\description{
Constructor to instantiate object(s) of class `TAIndicator`.
It calls the object class constructor with provided parameters.
}
\examples{
# Call the factory with technical indicator setup.
ta <- taFactory(label = "ADX", method = "TTR::ADX",
  columns = list(HLC = c("High", "Low", "Close")),
  args = list(n = 10, maType = "EMA")
)
}
