% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\docType{class}
\name{TDIConnection-class}
\alias{TDIConnection-class}
\title{TDIConnection class}
\usage{
\S4method{initialize}{TDIConnection}(.Object, ...)
}
\description{
Wraps an object of class [TDIDriver-class] and the connection parameters.
Its purpose is to establish a connection and to store all connection details.

Connection(s) are established by the `TDIConnector`,
which keeps a connection pool to re-use connection(s).

Initialize object(s) of class `TDIConnection`.
}
\section{Slots}{

\describe{
\item{\code{.drv}}{Object of class [TDIDriver-class].}

\item{\code{.conn_args}}{List with connection parameters.}

\item{\code{.endpoints}}{List with available API endpoints.}
}}

\seealso{
Other TDI classes: 
\code{\link{TDIConnector-class}},
\code{\link{TDIDriver-class}},
\code{\link{TDIObject-class}}

Other TDIConnection generics: 
\code{\link{getSymbol}()}

Other TDI classes: 
\code{\link{TDIConnector-class}},
\code{\link{TDIDriver-class}},
\code{\link{TDIObject-class}}
}
\concept{TDI classes}
\concept{TDIConnection generics}
