% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\name{TDIConnection}
\alias{TDIConnection}
\title{TDIConnection (R6 class constructor)}
\description{
`TDIConnection` class is a generic and abstract R6 class,
that wraps an object of class `TDIDriver` with the connection parameters.
Its purpose is to provide the capabilities to connect with the API.
Each API extends this class to implement the API specific capabilities.
}
\section{Super class}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{TDIConnection}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{driver}}{An object of class `TDIDriver`.}

\item{\code{conn_args}}{List with connection parameters.}

\item{\code{endpoints}}{List with available API endpoints.}

\item{\code{values}}{List collection with valid parameter values.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TDIConnection$new()}}
\item \href{#method-getInstrument}{\code{TDIConnection$getInstrument()}}
\item \href{#method-getChart}{\code{TDIConnection$getChart()}}
\item \href{#method-getCashFlow}{\code{TDIConnection$getCashFlow()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the API connection for the sub-class.
This class is abstract and cannot be instantiated.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIConnection$new(driver, conn_args, endpoints, values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{driver}}{An object of class `TDIDriver`.}

\item{\code{conn_args}}{List with connection parameters.}

\item{\code{endpoints}}{List with available API endpoints.}

\item{\code{values}}{List with allowed values for parameter(s).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `TDIConnection`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInstrument"></a>}}
\if{latex}{\out{\hypertarget{method-getInstrument}{}}}
\subsection{Method \code{getInstrument()}}{
Interface method to be implemented by sub-class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIConnection$getInstrument(symbol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Symbol to identify the instrument.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the calling arguments.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getChart"></a>}}
\if{latex}{\out{\hypertarget{method-getChart}{}}}
\subsection{Method \code{getChart()}}{
Interface method to be implemented by sub-class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIConnection$getChart(
  symbol,
  range = NULL,
  from = NULL,
  to = NULL,
  interval = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Symbol to identify the instrument.}

\item{\code{range}}{Optional period range.}

\item{\code{from}}{Optional start date of period.}

\item{\code{to}}{Optional end date of period.}

\item{\code{interval}}{Optional interval.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the calling arguments.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCashFlow"></a>}}
\if{latex}{\out{\hypertarget{method-getCashFlow}{}}}
\subsection{Method \code{getCashFlow()}}{
Interface method to be implemented by sub-class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIConnection$getCashFlow(symbol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Symbol to identify the instrument.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the calling arguments.
}
}
}
