% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIDriver.R
\docType{class}
\name{TDIDriver-class}
\alias{TDIDriver-class}
\alias{getInfo,TDIDriver-method}
\title{TDIDriver class}
\usage{
\S4method{getInfo}{TDIDriver}(obj, ...)
}
\description{
Base class for all TDI drivers (e.g., Yahoo, IEX, etc).
The virtual class `TDIDriver` is the generic class for API specific driver(s).

Each API implements an inherited [TDIDriver-class] to create the API driver object.
Object(s) of classes inheriting from [TDIDriver-class] to access the API.
}
\seealso{
Other TDI classes: 
\code{\link{TDIConnection-class}},
\code{\link{TDIConnector-class}},
\code{\link{TDIObject-class}},
\code{\link{TDIResult-class}}

Other TDIDriver generics: 
\code{\link{apiConnect}()},
\code{\link{getInfo}()}
}
\concept{TDI classes}
\concept{TDIDriver generics}
