% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIObject.R
\docType{class}
\name{TDIObject-class}
\alias{TDIObject-class}
\title{TDIObject class}
\description{
Base class for all TDI classes (e.g., drivers, etc). This
is a virtual Class: No objects may be created from it.

More generally, the TDI defines a small set of classes and generics to
facilitate a common interface for trading data.
}
\details{
Trading Data Interface (TDI)

Implement TDI following concept from RDI.

Main classes:
- TDIObject, a common base class
- TDIDriver, the instrument data source/driver
- TDIResult, results returned by API(s)
- TDIInstrument, represents generic instrument
- TDIIndex, represents specific instrument
- TDI<type>, specific instrument types
- TDIExchange, represent the exchange
}
\seealso{
Inspired by \link[DBI]{DBIObject}

Other TDI classes: 
\code{\link{TDIConnection-class}},
\code{\link{TDIConnector-class}},
\code{\link{TDIDriver-class}},
\code{\link{TDIResult-class}}
}
\concept{TDI classes}
