% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIResult.R
\docType{class}
\name{TDIResult-class}
\alias{TDIResult-class}
\title{TDIResult class}
\description{
Base class for all TDI result(s) sub-classes (e.g. Instrument, Index).
The virtual class `TDIResult` is the generic class for specific result(s) classes.
Each type of results implements a specific [TDIResult-class].
}
\seealso{
Other TDI classes: 
\code{\link{TDIConnection-class}},
\code{\link{TDIConnector-class}},
\code{\link{TDIDriver-class}},
\code{\link{TDIObject-class}}
}
\concept{TDI classes}
\concept{TDIResult generics}
