% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIResult.R
\docType{class}
\name{TDIResult-class}
\alias{TDIResult-class}
\title{TDIResult class}
\description{
Base class for all TDI result(s) sub-classes (e.g. Instrument, Index).
The class `TDIResult` is the generic class for specific result(s) classes,
and implements its shared generic methods.
}
\section{Slots}{

\describe{
\item{\code{.sources}}{Named list of sources for the data.}

\item{\code{.symbol}}{Unique identification symbol.}

\item{\code{.series}}{Time-series of class xts.}
}}

\seealso{
Other TDI classes: 
\code{\link{Instrument-class}},
\code{\link{TDIConnection-class}},
\code{\link{TDIConnector-class}},
\code{\link{TDIDriver-class}},
\code{\link{TDIIndicator-class}},
\code{\link{TDIObject-class}}

Other TDIResult generics: 
\code{\link{addSeries}()},
\code{\link{getSeries}()},
\code{\link{getSymbol}()},
\code{\link{setSeries}()}
}
\concept{TDI classes}
\concept{TDIResult generics}
