% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIResult.R
\name{TDIResult}
\alias{TDIResult}
\title{TDIResult (R6 class constructor)}
\description{
Generic base class for TDI result(s) sub-classes (e.g. TDIInstrument).
}
\details{
The TDI result(s) base class provides generic methods for results.
Validation methods allow for a quick check of API implementation classes.
These are recommended to be included in the tests for each API.
The validation check assumes that properties with NA are optional.
}
\section{Super class}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{TDIResult}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fields}{\code{TDIResult$fields()}}
\item \href{#method-validate}{\code{TDIResult$validate()}}
\item \href{#method-isValid}{\code{TDIResult$isValid()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="baseTDI" data-id="initialize">}\href{../../TDI/html/baseTDI.html#method-initialize}{\code{TDI::baseTDI$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fields"></a>}}
\if{latex}{\out{\hypertarget{method-fields}{}}}
\subsection{Method \code{fields()}}{
Return the properties of the object as a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIResult$fields(prop = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prop}}{Optional property in recursive call.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of object properties.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the properties of the object.
It is assumed that all NA properties are optional.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIResult$validate(prop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prop}}{Optional property in recursive call.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector with validation result.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-isValid}{}}}
\subsection{Method \code{isValid()}}{
Check if the object properties are valid.
A simple validation to check API implementation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TDIResult$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean as validation result.
}
}
}
