% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\name{validInterval}
\alias{validInterval}
\title{Validate the API interval parameter
Generic validation method for API interval parameter.}
\usage{
validInterval(obj, interval = NULL)
}
\arguments{
\item{obj}{An object of [TDIConnection-class].}

\item{interval}{Interval query parameter.}
}
\description{
Validate the API interval parameter
Generic validation method for API interval parameter.
}
\seealso{
Other TDIConnection generics: 
\code{\link{TDIConnection-class}},
\code{\link{getSymbols}()},
\code{\link{request}()},
\code{\link{validRange}()}
}
\concept{TDIConnection generics}
