% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDIConnection.R
\docType{methods}
\name{validRange}
\alias{validRange}
\title{Validate the API range parameter}
\usage{
validRange(obj, range = NULL)
}
\arguments{
\item{obj}{An object of class `TDIConnection`.}

\item{range}{Range query parameter.}
}
\description{
Generic validation method for API range parameter.
}
\seealso{
Other TDIConnection generics: 
\code{\link{TDIConnection-class}},
\code{\link{getChart}()},
\code{\link{request}()},
\code{\link{validInterval}()}
}
\concept{TDIConnection generics}
