% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahooAPI.R
\docType{class}
\name{YahooAPI-class}
\alias{YahooAPI-class}
\alias{initialize,YahooAPI-method}
\alias{request,YahooAPI-method}
\alias{validRange,YahooAPI-method}
\alias{validInterval,YahooAPI-method}
\alias{getSymbol,YahooAPI-method}
\title{Yahoo Finance API.}
\usage{
\S4method{initialize}{YahooAPI}(.Object, ...)

\S4method{request}{YahooAPI}(obj, path, query = NULL)

\S4method{validRange}{YahooAPI}(obj, range = NULL)

\S4method{validInterval}{YahooAPI}(obj, interval = NULL)

\S4method{getSymbol}{YahooAPI}(obj, symbol, range, from, to, interval)
}
\description{
Yahoo Finance API class inheriting from [TDIConnection-class].
This class implements the Yahoo Finance API requests.
}
\section{Slots}{

\describe{
\item{\code{.handle}}{Curl connection handle.}
}}

\keyword{internal}
