% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahooAPI.R
\docType{class}
\name{YahooAPI-class}
\alias{YahooAPI-class}
\alias{getSymbol,YahooAPI-method}
\title{Yahoo Finance API.}
\usage{
\S4method{initialize}{YahooAPI}(.Object, ...)

\S4method{getSymbol}{YahooAPI}(obj, symbol, range, from, to, interval = obj@.conn_args$interval)
}
\arguments{
\item{range}{A period range.}

\item{from}{Start date of period.}

\item{to}{End date of period.}

\item{interval}{Interval to retrieve quotes.}
}
\description{
Yahoo Finance API class inheriting from [TDIConnection-class].
This class implements the Yahoo Finance API requests.

Initialize object(s) of class `YahooAPI`.
}
\section{Slots}{

\describe{
\item{\code{.handle}}{Curl connection handle.}
}}

\keyword{internal}
