% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahooAPI.R
\name{yahooAPI}
\alias{yahooAPI}
\title{Yahoo Finance API}
\description{
Yahoo Finance API class inherits `TDIConnection` to implement API requests.
}
\details{
Yahoo Finance API provides a wide range of data that can be accessed via the below URL.
The data returned is specified by the requested modules in the query string.

https://query2.finance.yahoo.com/v10/finance/quoteSummary/MSFT?modules=assetProfile%2CsummaryProfile

Available modules:
- 'assetProfile',
- 'summaryProfile',
- 'summaryDetail',
- 'esgScores',
- 'price',
- 'incomeStatementHistory',
- 'incomeStatementHistoryQuarterly',
- 'balanceSheetHistory',
- 'balanceSheetHistoryQuarterly',
- 'cashflowStatementHistory',
- 'cashflowStatementHistoryQuarterly',
- 'defaultKeyStatistics',
- 'financialData',
- 'calendarEvents',
- 'secFilings',
- 'recommendationTrend',
- 'upgradeDowngradeHistory',
- 'institutionOwnership',
- 'fundOwnership',
- 'majorDirectHolders',
- 'majorHoldersBreakdown',
- 'insiderTransactions',
- 'insiderHolders',
- 'netSharePurchaseActivity',
- 'earnings',
- 'earningsHistory',
- 'earningsTrend',
- 'industryTrend',
- 'indexTrend',
- 'sectorTrend'
}
\section{Super classes}{
\code{\link[TDI:baseTDI]{TDI::baseTDI}} -> \code{\link[TDI:TDIConnection]{TDI::TDIConnection}} -> \code{yahooAPI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-getInstrument}{\code{yahooAPI$getInstrument()}}
\item \href{#method-getCashFlow}{\code{yahooAPI$getCashFlow()}}
\item \href{#method-getChart}{\code{yahooAPI$getChart()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TDI" data-topic="TDIConnection" data-id="initialize">}\href{../../TDI/html/TDIConnection.html#method-initialize}{\code{TDI::TDIConnection$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInstrument"></a>}}
\if{latex}{\out{\hypertarget{method-getInstrument}{}}}
\subsection{Method \code{getInstrument()}}{
Retrieve instrument details for the symbol.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{yahooAPI$getInstrument(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{see \code{\link{TDIConnection}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `Instrument` with details.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCashFlow"></a>}}
\if{latex}{\out{\hypertarget{method-getCashFlow}{}}}
\subsection{Method \code{getCashFlow()}}{
Retrieve cash flow data for the symbol.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{yahooAPI$getCashFlow(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{see \code{\link{TDIConnection}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `Instrument` with details.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getChart"></a>}}
\if{latex}{\out{\hypertarget{method-getChart}{}}}
\subsection{Method \code{getChart()}}{
Retrieve historical prices for the symbol.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{yahooAPI$getChart(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{see \code{\link{TDIConnection}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class `Instrument` with historical prices.
}
}
}
