% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahooDriver.R
\docType{class}
\name{YahooAPI}
\alias{YahooAPI}
\alias{yahoo-class}
\alias{apiConnect,yahoo-method}
\alias{getInfo,yahoo-method}
\title{Yahoo Finance API driver}
\usage{
YahooAPI()

\S4method{apiConnect}{yahoo}(obj, ...)

\S4method{getInfo}{yahoo}(obj, ...)
}
\description{
Driver for TDI object(s) to facilitatie access the Yahoo Finance API.

Driver class to implement Yahoo Finance API access.

The driver class `yahoo` is named with commonly used source name.
This makes is simpler to keep references to the driver with the data.
}
\keyword{internal}
