% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahooDriver.R
\docType{class}
\name{YahooDriver}
\alias{YahooDriver}
\alias{yahoo-class}
\alias{apiConnect,yahoo-method}
\title{Yahoo Finance API driver}
\usage{
YahooDriver()

\S4method{apiConnect}{yahoo}(obj, ...)
}
\description{
Helper function to establish a connection with Yahoo Finance API.

Driver class to implement Yahoo Finance API access.

The driver class `yahoo` is named with commonly used source name.
This makes is simpler to keep references to the driver with the data.
}
\keyword{internal}
